+++ src/App.tsx
@@ -3,13 +3,13 @@ import { Route, Routes } from 'react-router-dom'
 import './App.css'
 import PageOfflineGame from './pages/offline-game/PageOfflineGame'
 import PageUserHome from './pages/user-home/PageUserHome'
-import { PAGES } from './constants/pages'
+import { PAGES } from './utils/constants/pages'
 import PagePlayWithFriend from './pages/play-with-friend/PagePlayWithFriend'
-import { ToastContextProvider } from './contexts/ToastContext'
-import { AlertContextProvider } from './contexts/AlertContext'
+import { ToastContextProvider } from './utils/contexts/ToastContext'
+import { AlertContextProvider } from './utils/contexts/AlertContext'
 import PageExtras from './pages/extras/PageExtras'
 import PageHowToPlay from './pages/how-to-play/PageHowToPlay'
-import { BackDropContextProvider } from './contexts/BackDropContext'
+import { BackDropContextProvider } from './utils/contexts/BackDropContext'
 import PageAboutUs from './pages/extras/about-us/PageAboutUs'
 
 function App () {
