+++ src/components/menus/main-menu/MainMenu.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { PAGES } from '../../../constants/pages'
+import { PAGES } from '../../../utils/constants/pages'
 import MenuButton from '../../buttons/menu-button/MenuButton'
 
 const MainMenu: React.FunctionComponent = () => {
