+++ src/react-app-env.d.ts
@@ -1,1 +1,1 @@
-declare module '*.mp3'; // used for importing mp3 files
+declare module '*.mp3' // used for importing mp3 files
